mtd: spi-nor: implement OTP support for Macronix flashes
Use the new OTP ops to implement OTP access on Macronix flashes. The
Macronix flashes provides one OTP area which is either programmed with
an electonic serial number and locked by the factory or is empty and can
be locked by the user.

This was tested on a Macronix MX25L6405D.

Signed-off-by: Michael Walle <michael@walle.cc>