mtd: spi-nor: add support for MX25U3232F
This was tested on a custom board using the Freescale DSPI SPI
controller. Also locking was tested.

This flash uses 4 bits for block protection. Additionally, it
has TB bit, but this is only one-time-programmable. Thus we
ignore it for now until we have a good idea how to actually
program it.

To see the differences between MX25U3235F and MX25U3232F see the
application note AN0709 publicly available here:
http://www.macronix.com.tw/Lists/ApplicationNote/Attachments/2040/AN0709V1-The%20Comparison%20of%20MX25U3232F%20and%20MX25U3235F.pdf

Signed-off-by: Michael Walle <michael@walle.cc>