mtd: spi-nor: add support for OTP TB bit
Some flashes, for example the Macronix MX25U3232F has a TB bit to select
between top and bottom protection. But this bit is just one time
programmable. Thus we treat is as read-only in the current
locking/unlocking implementation.

That means that we have to (1) figure out whether the flash is top or
bottom protected (ie. the can_be_top/can_be_bottom setting depend on the
current TB setting and wether it is OTP or not) and (2) that we may only
write it back if it is not OTP.

It is out of the scope of this patch of how to actually program this
bit, but if we want to support these flashes with locking, we have to
at least handle this bit.

This patch already lays out the foundation to have the TB bit sepearate
from the status register, which also applies to the MX25U3232F.

Signed-off-by: Michael Walle <michael@walle.cc>