mtd: spi-nor: add support for TB bit in control register
Some flashes like the Macronix MX25U3232F has the TB bit in the control
register and not in the status register. Unfortunately, this means that
we now have to read two registers and pass its contents to various
functions. It also means that these functions have to be renamed,
because they are all ending in _sr() but they don't exclusivly operate
on the status register anymore.

Signed-off-by: Michael Walle <michael@walle.cc>