mtd: spi-nor: add sysfs support
Add support to read the raw register values as well as the SFDP table.
Raw register values are useful for debugging SPI flash issues. Not all
flashes list their SFDP table contents in their datasheet. So having
that is useful, too. But it also helps in bug reports from users.

Signed-off-by: Michael Walle <michael@walle.cc>