arm64: dts: ls1028a: assign a static clock rate to the pixel PLL
Due to a hardware bug in the LS1028A it is not possible to switch the
VCO frequency of the PLL. Thus we cannot change the pixel clock freely
but have to preset it to the correct frequency which fits the desired
resolution. It is possible program the output divider of the PLL during
runtime. Therefore it is possible to cover both 1920x1080 and 3840x2160
because both can use the same VCO frequency.

Signed-off-by: Michael Walle <michael@walle.cc>