arm64: dts: ls1028a: sl28: enable OTG on usb0
The sl28 base module which follows the SMARC specifications routes usb0
to USB OTG pins and provides the OTG_ID pin for that port as well. Thus
from the module's point of view this port is an OTG port.

A carrier may then to choose if it implements the OTG part, that is a
Micro-AB connector or a dual-role USB-C port, or if it just provides an
A connector. In the latter case, the OTG_ID pin should be connected to
ground. Unfortunatly, most carriers doesn't really care about it, which
leaves the port in device mode; specifically, the KBox-A-230-LS as well
as early revisions of the ADS2 carrier don't connect the OTG_ID pin to
ground. Thus hardwire the usb0 port on these carriers to host mode.

Signed-off-by: Michael Walle <michael@walle.cc>