mtd: spi-nor: debugfs: expose status registers
For now, to verify some features of the flash, we relied on indirect
methods. For example, to verify whether the block protection worked
correctly, we protected the region and then expected that this region
wasn't writable nor erasebale anymore.

By exposing the status registers, we can now see whether the block
protection bits are set correctly by the lock and unlock ioctls. There
is no need to write, read or erase the flash anymore.

Signed-off-by: Michael Walle <michael@walle.cc>