net: phy: introduce phy_supports_c45_transfers()
Provide a helper to determine if a PHY supports access to its C45
register space. For now, this is true if both the bus supports C45
transfers and the PHY also responds to them.
This is a preparation patch to remove the is_c45 field. No functional
change.

Signed-off-by: Michael Walle <mwalle@kernel.org>
---
v4:
 - rename to phy_supports_c45_transfers()

v3:
 - rename to phy_has_c45_registers()