net: phy: replace is_c45 with phy_transfer_mode
Instead of tracing whether the PHY is a C45 one, use the method how the
PHY is accessed. For now, that is either by C22 or by C45 transactions.
There is no functional change, just a semantical difference.

This is a preparation patch to add a third access method C45-over-C22.

Signed-off-by: Michael Walle <mwalle@kernel.org>
---
v4:
 - rebase to latest net-next
 - rename to phy_transfer_mode

v3:
 - use 22 and 45 as constants for the enum to catch errors
 - move enum in struct next to enum phy_state