net: phy: simplify comparing C45 PHY IDs
Everytime the a PHY C45 ID is compared to a given value, it is first
checked whether the PHY was probed via by C45 accesses. This is
unnecessary because the IDs are initialized with all-ones. Thus, only
if the PHY has been discovered via C45, the ID will be different than
the default all-one value.

Drop the unnecessary .is_c45 test. This is a preparation patch to remove
the .is_c45 property.

Signed-off-by: Michael Walle <mwalle@kernel.org>
---
v4:
 - new patch