dt-bindings: display: bridge: tc358775: make stby gpio optional
For a normal operation, the stby GPIO is needed. There are boards, where
these voltages are statically enabled during board power-up and thus,
the stby GPIO is also hard strapped to "always enabled".

The reset pin is required because once the PPI (PHY protocol interface)
is started, it can only be stopped by asserting the reset pin.

Signed-off-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Michael Walle <mwalle@kernel.org>